Renamed to teraslice-cli and move job all commands under cmds/job.
