+++ spec/errors-spec.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Promise = require('bluebird');
-const errors = require('../cmds/errors');
+const errors = require('../cmds/job/errors');
 
 const argv = {
     job_file: 'spec/fixtures/test_job_file.json'
