+++ spec/resume-spec.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Promise = require('bluebird');
-const resume = require('../cmds/resume');
+const resume = require('../cmds/job/resume');
 
 const argv = {
     job_file: 'spec/fixtures/test_job_file.json'
